refactor: introduce io/row, io/pattern, io/strategy subpackages
Moves 8 types from the io root into named subpackages to reduce clutter:
- io/row/: Row, ParsedRow, UnmatchedRow
- io/pattern/: FixedFormatMatchPattern, RegexFixedFormatMatchPattern
- io/strategy/: UnmatchStrategy, ParseErrorStrategy, PartialChunkStrategy

MultiMatchStrategy stays in io/ because its resolve() method references
ClassPatternMapping which is package-private in io/. All readers, writers,
and builders (which stay in io/ root) gain the necessary imports. Test file
TestRegexFixedFormatMatchPattern moves to the io/pattern/ test subpackage.
