+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/io/FixedFormatLineProcessor.java
@@ -17,6 +17,8 @@ package com.ancientprogramming.fixedformat4j.io;
 
 import com.ancientprogramming.fixedformat4j.exception.FixedFormatException;
 import com.ancientprogramming.fixedformat4j.format.FixedFormatManager;
+import com.ancientprogramming.fixedformat4j.io.strategy.ParseErrorStrategy;
+import com.ancientprogramming.fixedformat4j.io.strategy.UnmatchStrategy;
 
 import java.util.List;
 import java.util.function.BiConsumer;
