+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/io/pattern/FixedFormatMatchPattern.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ancientprogramming.fixedformat4j.io;
+package com.ancientprogramming.fixedformat4j.io.pattern;
 
 /**
  * Strategy interface for deciding whether a line in a fixed-format file should be loaded
