+++ fixedformat4j/src/test/java/com/ancientprogramming/fixedformat4j/io/TestFixedFormatRoundTrip.java
@@ -1,6 +1,11 @@
 package com.ancientprogramming.fixedformat4j.io;
 
 import com.ancientprogramming.fixedformat4j.format.impl.FixedFormatManagerImpl;
+import com.ancientprogramming.fixedformat4j.io.pattern.FixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.pattern.RegexFixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.row.ParsedRow;
+import com.ancientprogramming.fixedformat4j.io.row.Row;
+import com.ancientprogramming.fixedformat4j.io.row.UnmatchedRow;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
