+++ fixedformat4j/src/test/java/com/ancientprogramming/fixedformat4j/io/TestFixedFormatRowReader.java
@@ -1,5 +1,10 @@
 package com.ancientprogramming.fixedformat4j.io;
 
+import com.ancientprogramming.fixedformat4j.io.pattern.FixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.pattern.RegexFixedFormatMatchPattern;
+import com.ancientprogramming.fixedformat4j.io.row.ParsedRow;
+import com.ancientprogramming.fixedformat4j.io.row.Row;
+import com.ancientprogramming.fixedformat4j.io.row.UnmatchedRow;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
